Changes due to code review
Adding a factory class for creating instances of nexus_xxx_remote
Change to kind_of? in tasks.rb because I was actually doing it wrong
