+++ lib/nexus_cli.rb
@@ -2,7 +2,7 @@ require 'nexus_cli/tasks'
 require 'nexus_cli/cli'
 require 'nexus_cli/errors'
 require 'nexus_cli/kernel'
-require 'nexus_cli/nexus_remote'
+require 'nexus_cli/nexus_remote_factory'
 require 'nexus_cli/nexus_oss_remote'
 require 'nexus_cli/nexus_pro_remote'
